ai_backend/requirements.txt
fastapi[standard]
g4f
python-dotenv